bugfix for ckg internal server error due to not fixed Jinja dependency: upgrade flask (1.x -> 2.x) and dash (1.x -> 2.x) and adapt ckg code accordingly (use new dash pages feature and remove boilerplate navigation code)
