+++ ckg/ckg_utils.py
@@ -3,6 +3,7 @@ import numpy as np
 import yaml
 import json
 import logging
+from pathlib import Path
 
 
 def read_ckg_config(key=None):
@@ -154,3 +155,4 @@ class DictDFEncoder(json.JSONEncoder):
         if hasattr(obj, 'to_json'):
             return obj.to_json(orient='records')
         return json.JSONEncoder.default(self, obj)
+
