+++ ckg/config/ckg_config.yml
@@ -1,0 +1,23 @@
+version: 1.0
+ckg_directory: "/home/st/Schreibtisch/CKG/ckg"
+data_directory: "/home/st/Schreibtisch/CKG/data"
+archive_directory: "/home/st/Schreibtisch/CKG/data/archive"
+databases_directory: "/home/st/Schreibtisch/CKG/data/databases"
+experiments_directory: "/home/st/Schreibtisch/CKG/data/experiments"
+ontologies_directory: "/home/st/Schreibtisch/CKG/data/ontologies"
+users_directory: "/home/st/Schreibtisch/CKG/data/users"
+stats_directory: "/home/st/Schreibtisch/CKG/data/stats"
+downloads_directory: "/home/st/Schreibtisch/CKG/data/downloads"
+reports_directory: "/home/st/Schreibtisch/CKG/data/reports"
+tmp_directory: "/home/st/Schreibtisch/CKG/data/tmp"
+imports_directory: "/home/st/Schreibtisch/CKG/data/imports"
+imports_databases_directory: "/home/st/Schreibtisch/CKG/data/imports/databases"
+imports_experiments_directory: "/home/st/Schreibtisch/CKG/data/imports/experiments"
+imports_ontologies_directory: "/home/st/Schreibtisch/CKG/data/imports/ontologies"
+imports_users_directory: "/home/st/Schreibtisch/CKG/data/imports/users"
+imports_curated_directory: "/home/st/Schreibtisch/CKG/data/imports/curated"
+log_directory: "/home/st/Schreibtisch/CKG/log"
+report_manager_log: "/home/st/Schreibtisch/CKG/ckg/config/report_manager_log.config"
+graphdb_connector_log: "/home/st/Schreibtisch/CKG/ckg/config/graphdb_connector_log.config"
+graphdb_builder_log: "/home/st/Schreibtisch/CKG/ckg/config/graphdb_builder_log.config"
+analytics_factory_log: "/home/st/Schreibtisch/CKG/ckg/config/analytics_factory_log.config"
