+++ ckg/config/graphdb_builder_log.config
@@ -1,0 +1,45 @@
+{   "version": 1,
+    "disable_existing_loggers": false,
+    "formatters": {
+            "simple": {
+                "format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+            }
+    },
+    "handlers": {
+            "console": {
+                "class": "logging.StreamHandler",
+                "level": "NOTSET",
+                "formatter": "simple",
+                "stream": "ext://sys.stdout"
+            },
+            "info_file_handler": {
+                "class": "logging.handlers.RotatingFileHandler",
+                "level": "INFO",
+                "formatter": "simple",
+                "filename": "/home/st/Schreibtisch/CKG/log/graphdb_builder.log",
+                "maxBytes": 10485760,
+                "backupCount": 20,
+                "encoding": "utf8"
+            },
+            "error_file_handler": {
+                "class": "logging.handlers.SMTPHandler",
+                "level": "CRITICAL",
+                "formatter": "simple",
+                "mailhost": "localhost",
+                "fromaddr": "error@ckg.com",
+                "toaddrs": "alberto.santos@cpr.ku.dk",
+                "subject": "CKG ERROR - GraphDB builder"
+            }
+    },
+    "loggers": {
+            "my_module": {
+                "level": "ERROR",
+                "handlers": ["console"],
+                "propagate": "no"
+            }
+    },
+    "root":{
+        "level": "WARNING",
+        "handlers": ["info_file_handler", "error_file_handler"]
+    }
+}
