+++ ckg/config/report_manager_log.config
@@ -1,0 +1,45 @@
+{   "version": 1,
+    "disable_existing_loggers": false,
+    "formatters": {
+            "simple": {
+                "format": "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+            }
+    },
+    "handlers": {
+            "console": {
+                "class": "logging.StreamHandler",
+                "level": "NOTSET",
+                "formatter": "simple",
+                "stream": "ext://sys.stdout"
+            },
+            "info_file_handler": {
+                "class": "logging.handlers.RotatingFileHandler",
+                "level": "INFO",
+                "formatter": "simple",
+                "filename": "/home/st/Schreibtisch/CKG/log/report_manager.log",
+                "maxBytes": 10485760,
+                "backupCount": 20,
+                "encoding": "utf8"
+            },
+            "error_file_handler": {
+                "class": "logging.handlers.SMTPHandler",
+                "level": "CRITICAL",
+                "formatter": "simple",
+                "mailhost": "localhost",
+                "fromaddr": "error@ckg.com",
+                "toaddrs": "alberto.santos@cpr.ku.dk",
+                "subject": "CKG ERROR -  Report manager"
+            }
+    },
+    "loggers": {
+            "my_module": {
+                "level": "ERROR",
+                "handlers": ["console"],
+                "propagate": "no"
+            }
+    },
+    "root":{
+        "level": "INFO",
+        "handlers": ["info_file_handler", "error_file_handler"]
+    }
+}
