+++ ckg/report_manager/apps/dataUpload.py
@@ -1,13 +1,15 @@
 import os
-import sys
 import re
-import pandas as pd
+import sys
+
 import numpy as np
+import pandas as pd
+
 from ckg import ckg_utils
-from ckg.graphdb_connector import connector
-from ckg.graphdb_builder import builder_utils
-from ckg.graphdb_connector import query_utils
 from ckg.analytics_core.viz import viz
+from ckg.graphdb_builder import builder_utils
+from ckg.graphdb_connector import connector
+from ckg.graphdb_connector import query_utils
 
 ckg_config = ckg_utils.read_ckg_config()
 log_config = ckg_config['graphdb_builder_log']
@@ -28,7 +30,9 @@ def get_data_upload_queries():
     except Exception as err:
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Reading queries from file {}: {}, file: {},line: {}".format(err, queries_path, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Reading queries from file {}: {}, file: {},line: {}".format(err, queries_path, sys.exc_info(),
+                                                                                    fname, exc_tb.tb_lineno))
 
     return data_upload_cypher
 
@@ -53,7 +57,9 @@ def get_new_subject_identifier(driver):
         subject_identifier = None
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Getting new subject identifiers: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Getting new subject identifiers: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
     return subject_identifier
 
 
@@ -75,7 +81,9 @@ def get_new_biosample_identifier(driver):
         identifier = None
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Getting new biological sample identifiers: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Getting new biological sample identifiers: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     return identifier
 
@@ -97,7 +105,9 @@ def get_new_analytical_sample_identifier(driver):
         identifier = None
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Getting new analytical sample identifiers: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Getting new analytical sample identifiers: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     return identifier
 
@@ -119,13 +129,15 @@ def get_subjects_enrolled_in_project(driver, projectId):
         query = data_upload_cypher[query_name]['query']
         for q in query.split(';')[0:-1]:
             if '$' in q:
-                result = connector.getCursorData(driver, q+';', parameters={'external_id': str(projectId)})
+                result = connector.getCursorData(driver, q + ';', parameters={'external_id': str(projectId)})
             else:
-                result = connector.getCursorData(driver, q+';')
+                result = connector.getCursorData(driver, q + ';')
     except Exception as err:
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Getting new subjects enrolled in project: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Getting new subjects enrolled in project: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
     return result.values
 
 
@@ -142,7 +154,9 @@ def check_samples_in_project(driver, projectId):
     except Exception as err:
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Checking whether samples exist in project: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Checking whether samples exist in project: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     return result
 
@@ -160,7 +174,9 @@ def check_external_ids_in_db(driver, projectId):
     except Exception as err:
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Checking if external identifiers exist in the database: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Checking if external identifiers exist in the database: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     return result
 
@@ -177,11 +193,13 @@ def remove_samples_nodes_db(driver, projectId):
         project_cypher = ckg_utils.get_queries(os.path.join(directory, queries_path))
         query = project_cypher[query_name]['query'].replace('PROJECTID', projectId).split(';')[:-2]
         for q in query:
-            result = connector.commitQuery(driver, q+';')
+            result = connector.commitQuery(driver, q + ';')
     except Exception as err:
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Removing nodes associated to project: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Removing nodes associated to project: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     return result
 
@@ -197,7 +215,7 @@ def create_new_subjects(driver, data, projectId):
     subject_id = get_new_subject_identifier(driver)
     if subject_id is None:
         subject_id = '1'
-    subject_ids = ['S'+str(i) for i in np.arange(int(subject_id), int(subject_id) + len(external_ids))]
+    subject_ids = ['S' + str(i) for i in np.arange(int(subject_id), int(subject_id) + len(external_ids))]
     subject_dict = dict(zip(external_ids, subject_ids))
     query_name = 'create_project_subject'
     for external_id, subject_id in subject_dict.items():
@@ -207,11 +225,13 @@ def create_new_subjects(driver, data, projectId):
             data_upload_cypher = get_data_upload_queries()
             queries = data_upload_cypher[query_name]['query'].split(';')[:-1]
             for query in queries:
-                res = connector.commitQuery(driver, query+';', parameters=parameters)
+                res = connector.commitQuery(driver, query + ';', parameters=parameters)
         except Exception as err:
             exc_type, exc_obj, exc_tb = sys.exc_info()
             fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-            logger.error("Error: {}. Creating new subjects: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+            logger.error(
+                "Error: {}. Creating new subjects: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                    err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     data['subject id'] = data['subject external_id'].map(subject_dict)
 
@@ -231,23 +251,25 @@ def create_new_biosamples(driver, data):
     if biosample_id is None:
         biosample_id = '1'
 
-    biosample_ids = ['BS'+str(i) for i in np.arange(int(biosample_id), int(biosample_id) + len(external_ids))]
+    biosample_ids = ['BS' + str(i) for i in np.arange(int(biosample_id), int(biosample_id) + len(external_ids))]
     biosample_dict = dict(zip(external_ids, biosample_ids))
     biosample_subject_dict = dict(zip(external_ids, subject_ids))
     query_name = 'create_subject_biosamples'
     for external_id, biosample_id in biosample_dict.items():
         subject_id = biosample_subject_dict[external_id]
-        parameters = {'external_id': str(external_id), 'biosample_id':biosample_id, 'subject_id': subject_id}
+        parameters = {'external_id': str(external_id), 'biosample_id': biosample_id, 'subject_id': subject_id}
         try:
             query = ''
             data_upload_cypher = get_data_upload_queries()
             queries = data_upload_cypher[query_name]['query'].split(';')[:-1]
             for query in queries:
-                res = connector.commitQuery(driver, query+';', parameters=parameters)
+                res = connector.commitQuery(driver, query + ';', parameters=parameters)
         except Exception as err:
             exc_type, exc_obj, exc_tb = sys.exc_info()
             fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-            logger.error("Error: {}. Creating biological samples: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+            logger.error(
+                "Error: {}. Creating biological samples: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                    err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     data['biological_sample id'] = data['biological_sample external_id'].map(biosample_dict)
 
@@ -282,13 +304,16 @@ def create_new_ansamples(driver, data):
             data_upload_cypher = get_data_upload_queries()
             queries = data_upload_cypher[query_name]['query'].split(';')[:-1]
             for query in queries:
-                res = connector.commitQuery(driver, query+';', parameters=parameters)
+                res = connector.commitQuery(driver, query + ';', parameters=parameters)
         except Exception as err:
             exc_type, exc_obj, exc_tb = sys.exc_info()
             fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-            logger.error("Error: {}. Creating analytical samples: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+            logger.error(
+                "Error: {}. Creating analytical samples: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                    err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
-    data = data.rename(columns={'asample_id': 'analytical_sample id', 'external_id': 'analytical_sample external_id', 'biosample_id': 'biological_sample id'})
+    data = data.rename(columns={'asample_id': 'analytical_sample id', 'external_id': 'analytical_sample external_id',
+                                'biosample_id': 'biological_sample id'})
 
     return data
 
@@ -362,21 +387,26 @@ def get_project_information(driver, project_id):
             code = section['query']
             queries.extend(code.replace("PROJECTID", project_id).split(';')[0:-1])
         for query in queries:
-            result = connector.sendQuery(driver, query+";")[0]
+            result = connector.sendQuery(driver, query + ";")[0]
             data.append(result)
     except Exception as err:
         exc_type, exc_obj, exc_tb = sys.exc_info()
         fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-        logger.error("Error: {}. Creating analytical samples: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
+        logger.error(
+            "Error: {}. Creating analytical samples: Query name ({}) - Query ({}), error info: {}, file: {},line: {}".format(
+                err, query_name, query, sys.exc_info(), fname, exc_tb.tb_lineno))
 
     if data:
         for i, j in enumerate(data):
             df = pd.DataFrame([data[i]], columns=data[i].keys())
             header = '_'.join(df.columns[0].split('_', 1)[1:]).capitalize()
             df.rename(columns={df.columns[0]: 'project'}, inplace=True)
-            res.append(viz.get_table(df, identifier='new_project_{}'.format(header), args={'title':'{} data uploaded for project {}'.format(header, project_id)}))
+            res.append(viz.get_table(df, identifier='new_project_{}'.format(header),
+                                     args={'title': '{} data uploaded for project {}'.format(header, project_id)}))
     else:
         res = None
-        logger.error("Error: No data was uploaded for project: {}. Review your experimental design and data files and the logs for errors.".format(project_id))
+        logger.error(
+            "Error: No data was uploaded for project: {}. Review your experimental design and data files and the logs for errors.".format(
+                project_id))
 
     return res
