--- ckg/report_manager/apps/initialApp.py
@@ -1,3 +1,0 @@
-import dash_html_components as html
-
-layout =  [html.H1('Welcome to the app'), html.H3('You are successfully authorized')]
