+++ ckg/report_manager/utils.py
@@ -1,20 +1,19 @@
-import os
-import pandas as pd
-import dash_html_components as html
-import dash_core_components as dcc
-from ckg.graphdb_connector import connector
-from datetime import datetime
-import bs4 as bs
-import random
-import numpy as np
-import chart_studio.plotly as py
 import base64
-from xhtml2pdf import pisa
 import json
-from urllib import request
+import os
+import random
 import shutil
 import smtplib
+from datetime import datetime
 from email.message import EmailMessage
+from urllib import request
+
+import bs4 as bs
+import chart_studio.plotly as py
+from dash import dcc
+from dash import html
+from xhtml2pdf import pisa
+
 from ckg import ckg_utils
 
 
@@ -28,25 +27,26 @@ def send_message_to_slack_webhook(message, message_to, username='albsantosdel'):
     if os.path.exists(webhook_file):
         with open(webhook_file, 'r') as hf:
             webhook_url = hf.read()
-        post = {"text": "@{} : {}".format(message_to, message), "username": username, "icon_url": "https://slack.com/img/icons/app-57.png"}
+        post = {"text": "@{} : {}".format(message_to, message), "username": username,
+                "icon_url": "https://slack.com/img/icons/app-57.png"}
         try:
             json_data = json.dumps(post)
             req = request.Request(webhook_url,
-                                data=json_data.encode('ascii'),
-                                headers={'Content-Type': 'application/json'})
+                                  data=json_data.encode('ascii'),
+                                  headers={'Content-Type': 'application/json'})
             resp = request.urlopen(req)
         except Exception as em:
             print("EXCEPTION: " + str(em))
 
 
 def send_email(message, subject, message_from, message_to):
-    msg = EmailMessage()    
+    msg = EmailMessage()
     msg.set_content(message)
     msg['Subject'] = subject
     msg['From'] = message_from
     msg['to'] = message_to
 
-    s = smtplib.SMTP('localhost:1025') 
+    s = smtplib.SMTP('localhost:1025')
     s.send_message(msg)
     s.quit()
 
@@ -54,12 +54,14 @@ def send_email(message, subject, message_from, message_to):
 def compress_directory(name, directory, compression_format='zip'):
     try:
         if os.path.exists(directory) and os.path.isdir(directory):
-            if os.path.exists(directory+'.zip'):
-                os.remove(directory+'.zip')
+            if os.path.exists(directory + '.zip'):
+                os.remove(directory + '.zip')
             shutil.make_archive(name, compression_format, directory)
             shutil.rmtree(directory)
         else:
-            print("The directory {} failed. Exists?{} Is dir?{}".format(directory, os.path.exists(directory) and os.path.isdir(directory)))
+            print("The directory {} failed. Exists?{} Is dir?{}".format(directory,
+                                                                        os.path.exists(directory) and os.path.isdir(
+                                                                            directory)))
     except Exception as err:
         print("Could not compress file {} in directory {}. Error: {}".format(name, directory, err))
 
@@ -94,20 +96,20 @@ def get_image(figure, width, height):
 def parse_html(html_snippet):
     html_parsed = bs.BeautifulSoup(html_snippet)
 
-    return html_parsed    
+    return html_parsed
 
 
 def hex2rgb(color):
     hex = color.lstrip('#')
-    rgb = tuple(int(hex[i:i+2], 16) for i in (0, 2, 4))
+    rgb = tuple(int(hex[i:i + 2], 16) for i in (0, 2, 4))
     rgba = rgb + (0.6,)
     return rgba
 
 
 def getNumberText(num):
     numbers = ["zero", "one", "two", "three", "four", "five", "six", "seven", "eight",
-        "nine", "ten", "eleven", "twelve", "thirteen", "fourteen", "fifteen",
-        "sixteen", "seventeen", "eighteen", "nineteen"]
+               "nine", "ten", "eleven", "twelve", "thirteen", "fourteen", "fifteen",
+               "sixteen", "seventeen", "eighteen", "nineteen"]
     if len(numbers) > num:
         return numbers[num]
     else:
@@ -148,11 +150,11 @@ def convert_html_to_pdf(source_html, output_filename):
 
     # convert HTML to PDF
     pisa_status = pisa.CreatePDF(
-            source_html,                # the HTML to convert
-            dest=result_file)           # file handle to recieve result
+        source_html,  # the HTML to convert
+        dest=result_file)  # file handle to recieve result
 
     # close output file
-    result_file.close()                 # close output file
+    result_file.close()  # close output file
 
     # return True on success and False on errors
     return pisa_status.err
