+++ requirements.txt
@@ -1,60 +1,212 @@
-pandas==0.24.2
-Flask==1.0.3
-passlib==1.7.1
-neo4j==4.2.0
-PyYAML==5.1.1
-wget==3.2
-requests==2.22.0
-biopython==1.73
-obonet==0.2.5
-rarfile==3.1
-dash==1.2.0
-redis>=3.2.1
-matplotlib>=3.1.1
-scipy==1.4.1
-cyjupyter==0.2.0
-pyvis==0.1.7.0
-tzlocal==1.5.1
-webweb==0.0.37
-beautifulsoup4==4.7.1
-sklearn==0.0
-statsmodels==0.10.0
-umap-learn>=0.3.9
-pingouin>=0.3.10
-python-louvain==0.13
-snfpy==0.2.2
-kmapper==1.2.0
-natsort==6.0.0
-mpmath==1.2.1
-combat==0.2.0
-lifelines==0.23.1
-cvxpy==1.1.7
-wordcloud==1.8.1
-nltk==3.5
-dash_cytoscape>=0.1.1
-networkx==2.5
-chart-studio==1.0.0
-xhtml2pdf==0.2.4
-lxml==4.3.4
-celery==5.0.5
-cffi>=1.14.0
-rpy2==3.0.5; platform_system!="Windows"
-h5py==2.10.0
-jupyter>=1.0.0
-xlrd==1.2.0
-openpyxl==3.0.1
+alabaster==0.7.12
+amqp==5.1.1
+appdirs==1.4.4
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+async-timeout==4.0.2
+attrs==22.1.0
+autograd==1.5
+autograd-gamma==0.5.0
+Babel==2.10.3
+backcall==0.2.0
 bcrypt==3.1.7
-tables==3.6.1
-psutil>=5.6.6
-plotly==4.9.0
-kaleido==0.0.1
-gseapy==0.9.19
-numpy==1.19.5
+beautifulsoup4==4.7.1
+billiard==3.6.4.0
+biopython==1.73
+bioservices==1.10.1
+bleach==5.0.1
+Brotli==1.0.9
+cached-property==1.5.2
+cattrs==22.1.0
+celery==5.2.7
+certifi==2022.9.24
+cffi==1.15.1
+chardet==3.0.4
+chart-studio==1.0.0
+CKG==1.0.0
+click==8.1.3
+click-didyoumean==0.3.0
+click-plugins==1.1.1
+click-repl==0.2.0
+colorama==0.4.5
+colorlog==6.7.0
+combat==0.2.0
+cvxpy==1.1.7
+cycler==0.11.0
+cyjupyter==0.2.0
+dash==2.7.0
+dash-core-components==2.0.0
+dash-cytoscape==0.3.0
+dash-html-components==2.0.0
+dash-table==5.0.0
+debugpy==1.6.3
+decorator==5.1.1
+defusedxml==0.7.1
+Deprecated==1.2.13
+dnspython==2.2.1
 docutils==0.15.2
-sphinx==2.3.1
-rst2pdf==0.98
-nbsphinx>=0.8.0
-ipykernel>=5.4.3
+easydev==0.12.0
+ecos==2.0.10
+entrypoints==0.4
+et-xmlfile==1.1.0
 eventlet==0.25.1
+exceptiongroup==1.0.0rc9
+fastjsonschema==2.16.2
+Flask==2.2.2
+Flask-Compress==1.13
+fonttools==4.37.3
+future==0.18.2
+gevent==21.12.0
+greenlet==1.1.3
+grequests==0.6.0
+gseapy==0.9.19
+h5py==2.10.0
+html5lib==1.1
+idna==2.8
+imagesize==1.4.1
+importlib-metadata==4.12.0
+importlib-resources==5.9.0
+iniconfig==1.1.1
+ipykernel==6.16.0
+ipython==7.34.0
+ipython-genutils==0.2.0
+ipywidgets==8.0.2
 itsdangerous==2.0.1
+jdcal==1.4.1
+jedi==0.18.1
+Jinja2==3.1.2
+joblib==1.2.0
+jsonschema==4.16.0
+jupyter==1.0.0
+jupyter-console==6.4.4
+jupyter-core==4.11.1
+jupyter_client==7.3.5
+jupyterlab-pygments==0.2.2
+jupyterlab-widgets==3.0.3
+kaleido==0.0.1
+kiwisolver==1.4.4
+kmapper==1.2.0
+kombu==5.2.4
+lifelines==0.23.1
+littleutils==0.2.2
+llvmlite==0.39.1
+lxml==4.3.4
+MarkupSafe==2.1.1
+matplotlib==3.5.3
+matplotlib-inline==0.1.6
+mistune==0.8.4
+monotonic==1.6
+mpmath==1.2.1
+natsort==6.0.0
+nbclient==0.5.13
+nbconvert==6.4.5
+nbformat==5.6.1
+nbsphinx==0.8.9
+neo4j==4.2.0
+nest-asyncio==1.5.6
+networkx==2.5
+nltk==3.5
+notebook==6.4.12
+numba==0.56.2
+numexpr==2.8.3
+numpy==1.19.5
+obonet==0.2.5
+openpyxl==3.0.1
+osqp==0.6.2.post5
+outdated==0.2.1
+packaging==21.3
+pandas==0.24.2
+pandas-flavor==0.2.0
+pandocfilters==1.5.0
+parso==0.8.3
+passlib==1.7.1
+patsy==0.5.2
+pexpect==4.8.0
+pickleshare==0.7.5
+Pillow==9.2.0
+pingouin==0.3.12
+pkgutil_resolve_name==1.3.10
+plotly==5.10.0
+pluggy==1.0.0
+prometheus-client==0.14.1
+prompt-toolkit==3.0.31
+psutil==5.9.2
+ptyprocess==0.7.0
+py==1.11.0
+pycparser==2.21
+Pygments==2.13.0
+pynndescent==0.5.7
+pyparsing==3.0.9
+PyPDF2==2.11.0
+pyrsistent==0.18.1
+pytest==7.1.3
+python-dateutil==2.8.2
+python-louvain==0.13
+pytz==2022.2.1
+pyvis==0.1.7.0
+PyYAML==5.1.1
+pyzmq==24.0.1
+qdldl==0.1.5.post2
+qtconsole==5.3.2
+QtPy==2.2.0
+rarfile==3.1
+redis==4.3.4
+regex==2022.9.13
 reportlab==3.6.6
+requests==2.22.0
+requests-cache==0.9.6
+retrying==1.3.3
+rpy2==3.0.5
+rst2pdf==0.98
+scikit-learn==1.0.2
+scipy==1.4.1
+scs==3.2.0
+seaborn==0.11.2
+Send2Trash==1.8.0
+simplegeneric==0.8.1
+six==1.16.0
+sklearn==0.0
+smartypants==2.0.1
+snfpy==0.2.2
+snowballstemmer==2.2.0
+soupsieve==2.3.2.post1
+Sphinx==2.3.1
+sphinxcontrib-applehelp==1.0.2
+sphinxcontrib-devhelp==1.0.2
+sphinxcontrib-htmlhelp==2.0.0
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.3
+sphinxcontrib-serializinghtml==1.1.5
+statsmodels==0.10.0
+suds-community==1.1.2
+tables==3.6.1
+tabulate==0.8.10
+tenacity==8.1.0
+terminado==0.16.0
+testpath==0.6.0
+threadpoolctl==3.1.0
+tomli==2.0.1
+tornado==6.2
+tqdm==4.64.1
+traitlets==5.4.0
+typing_extensions==4.3.0
+tzlocal==1.5.1
+umap-learn==0.5.3
+url-normalize==1.4.3
+urllib3==1.25.11
+vine==5.0.0
+wcwidth==0.2.5
+webencodings==0.5.1
+webweb==0.0.37
+Werkzeug==2.2.2
+wget==3.2
+widgetsnbextension==4.0.3
+wordcloud==1.8.1
+wrapt==1.14.1
+xarray==0.14.1
+xhtml2pdf==0.2.4
+xlrd==1.2.0
+xmltodict==0.13.0
+zipp==3.8.1
+zope.event==4.5.0
+zope.interface==5.4.0
